erros everywhere

+++ PokeLibrary 2.0/src/java/model/Preco.java
@@ -24,7 +24,7 @@ public class Preco implements Serializable{
     private Integer id;
     
     @Column(name = "PRECO_VERSAO", nullable = false)
-    private int preco;
+    private double preco;
     
     @ManyToMany(mappedBy = "preco")
     private List<Item> itens;
@@ -32,15 +32,15 @@ public class Preco implements Serializable{
     public Preco() {
     }
     
-    public Preco(int preco)
+    public Preco(double preco)
     {
         this.preco=preco;
     }
-    public int getPreco()
+    public double getPreco()
     {
         return this.preco;
     }
-    public void setPreco(int preco)
+    public void setPreco(double preco)
     {
         this.preco=preco;
     }
